move this example to urls.py
