+++ tutorial/urls.py
@@ -172,19 +172,143 @@ function. A few notes:
 - Each page can have interactive elements even though those elements may not
 be in the initial view. Dash handles these "dynamically generated" components
 gracefully: as they are rendered, they will trigger the
-callbacks with their initial values.'''),
-    html.Ul(html.Li(['''Since we're adding callbacks to elements that don't exist in the app.layout,
+callbacks with their initial values.
+- Since we're adding callbacks to elements that don't exist in the app.layout,
 Dash will raise an exception to warn us that we might be doing something
 wrong.  In this case, we're adding the elements through a callback, so we can
 ignore the exception by setting `app.config.suppress_callback_exceptions = True`.
 It is also possible to do this without suppressing callback exceptions. See the example
-"Dynamically Create a Layout for Multi-Page App Validation" in ''',
-    dcc.Link('Live Updates', href='/live-updates'), 'for more details.'])),
-    dcc.Markdown('''
+below for details.
 - You can modify this example to import the different page's `layout`s in different files.
 - This Dash Userguide that you're looking at is itself a multi-page Dash app, using 
 these same principles.
 '''),
+
+          dcc.Markdown('''
+
+***
+
+## Dynamically Create a Layout for Multi-Page App Validation
+
+Dash applies validation to your callbacks, which performs checks such as 
+validating the types of callback arguments and checking to see whether the 
+specified Input and Output components actually have the specified properties. 
+
+For full validation, all components within your callback must therefore appear 
+in the initial layout of your app, and you will see an error if they do not.
+However, in the case of more complex Dash apps that involve dynamic modification 
+of the layout (such as multi-page apps), not every component appearing in your 
+callbacks will be included in the initial layout.
+
+Since this validation is done before flask has any request context, you can create 
+a layout function that checks `flask.has_request_context()` and returns a complete
+layout to the validator if there is no request context and returns the incomplete
+index layout otherwise.
+'''),
+    dcc.SyntaxHighlighter(
+        '''
+import dash
+import dash_core_components as dcc
+import dash_html_components as html
+from dash.dependencies import Input, Output, State
+
+import flask
+
+app = dash.Dash(
+    __name__, 
+    external_stylesheets=['https://codepen.io/chriddyp/pen/bWLwgP.css']
+)
+
+url_bar_and_content_div = html.Div([
+    dcc.Location(id='url', refresh=False),
+    html.Div(id='page-content')
+])
+
+layout_index = html.Div([
+    dcc.Link('Navigate to "/page-1"', href='/page-1'),
+    html.Br(),
+    dcc.Link('Navigate to "/page-2"', href='/page-2'),
+])
+
+layout_page_1 = html.Div([
+    html.H2('Page 1'),
+    dcc.Input(id='input-1-state', type='text', value='Montreal'),
+    dcc.Input(id='input-2-state', type='text', value='Canada'),
+    html.Button(id='submit-button', n_clicks=0, children='Submit'),
+    html.Div(id='output-state'),
+    html.Br(),
+    dcc.Link('Navigate to "/"', href='/'),
+    html.Br(),
+    dcc.Link('Navigate to "/page-2"', href='/page-2'),
+])
+
+layout_page_2 = html.Div([
+    html.H2('Page 2'),
+    dcc.Dropdown(
+        id='page-2-dropdown',
+        options=[{'label': i, 'value': i} for i in ['LA', 'NYC', 'MTL']],
+        value='LA'
+    ),
+    html.Div(id='page-2-display-value'),
+    html.Br(),
+    dcc.Link('Navigate to "/"', href='/'),
+    html.Br(),
+    dcc.Link('Navigate to "/page-1"', href='/page-1'),
+])
+
+
+def serve_layout():
+    if flask.has_request_context():
+        return url_bar_and_content_div
+    return html.Div([
+        url_bar_and_content_div,
+        layout_index,
+        layout_page_1,
+        layout_page_2,
+    ])
+
+
+app.layout = serve_layout
+
+
+# Index callbacks
+@app.callback(Output('page-content', 'children'),
+              [Input('url', 'pathname')])
+def display_page(pathname):
+    if pathname == "/page-1":
+        return layout_page_1
+    elif pathname == "/page-2":
+        return layout_page_2
+    else:
+        return layout_index
+
+
+# Page 1 callbacks
+@app.callback(Output('output-state', 'children'),
+              [Input('submit-button', 'n_clicks')],
+              [State('input-1-state', 'value'),
+               State('input-2-state', 'value')])
+def update_output(n_clicks, input1, input2):
+    return ('The Button has been pressed {} times,'
+            'Input 1 is "{}",'
+            'and Input 2 is "{}"').format(n_clicks, input1, input2)
+
+
+# Page 2 callbacks
+@app.callback(Output('page-2-display-value', 'children'),
+              [Input('page-2-dropdown', 'value')])
+def display_value(value):
+    print('display_value')
+    return 'You have selected "{}"'.format(value)
+
+
+if __name__ == '__main__':
+    app.run_server(debug=True)
+    
+        ''',
+        language='python',
+        customStyle={'borderLeft': 'thin solid lightgrey'}
+    ),
 
           dcc.Markdown('''***
